moving project logic to selectors
