+++ web/src/lib/api.js
@@ -1,5 +1,4 @@
 import endpoints from '../endpoints'
-import { v4 as uuidV4 } from 'uuid'
 
 export function postLogin(username, password) {
 
@@ -53,7 +52,7 @@ export function getProjects(token) {
         const { Id, Name } = item;
         return { Id, Name };
       });
-      
+
       resolve(projectData);
     })
     .catch(e => {
@@ -87,54 +86,8 @@ export function getResources(token, { weekstart, weekend }) {
     })
     .then(handleErrors)
     .then(response => response.json())
-    .then(data => {
-
-      //collect all resource hours by projectId
-      const ProjectResourceHoursHashMap = data.reduce((accumulator, currentValue) => {
-
-        const { Project__c, Week_Start__c } = currentValue
-
-        return {
-          ...accumulator,
-          [ Project__c ]: {
-            ...accumulator[Project__c],
-            [ Week_Start__c ]: currentValue
-          }
-        }
-      }, {});
-
-      //create a Project Hashmap
-      const ProjectHashMap = data.reduce = data.reduce((accumulator, currentValue) => {
-        const { Project__c, Project__r } = currentValue
-
-        return {
-          ...accumulator,
-          [ Project__c ] : {
-            ...Project__r
-          }
-        }
-      }, {})
-
-
-      //generate final structure
-      const projectData = Object.values(ProjectHashMap).reduce((accumulator, currentValue) => {
-        const { Id, Name } = currentValue
-        const uuid = uuidV4();
-
-        return {
-          ...accumulator,
-          [ uuid ] : {
-            uuid,
-            Id,
-            Name,
-            values: {
-              ...ProjectResourceHoursHashMap[Id]
-            }
-          }
-        }
-      }, {});
-      resolve(projectData)
-    }).catch(e => {
+    .then(resolve)
+    .catch(e => {
       reject(new Error('Error processing get resources call'))
     })
   })
